requirements.txt
fastapi
uvicorn
pydantic[email]
sqlalchemy
psycopg2-binary
python-dotenv
pytest
pytest-cov
pytest-asyncio
coverage[toml]
httpx
sqlmodel
python-jose[cryptography]
boto3
python-multipart
authlib
itsdangerous
openai
numpy==1.26.4
chromadb==0.4.15
tiktoken
jinja2
bcrypt==4.0.1
passlib[bcrypt]==1.7.4